+++ CHANGELOG.md
@@ -11,6 +11,11 @@
 - `PostgreSQLResult` and `PostgreSQLResultRow` as the return value of a query.
   - Returned lists are protected with `UnmodifiableListView`.
 - `PostgreSQLConnection` and `_TransactionProxy` share the OID cache.
+- default value for `query(allowReuse = true)` is set only in the implementation method.
+
+**Breaking behaviour**
+
+- Table OIDs are always resolved to table names (and not only with mapped queries).
 
 ## 1.0.2
 - Add connection queue size